+++ src/main/java/com/universestay/project/admin/controller/UserManagementController.java
@@ -1,7 +1,7 @@
 package com.universestay.project.admin.controller;
 
-import com.universestay.project.admin.dto.PageHandler;
-import com.universestay.project.admin.dto.SearchCondition;
+import com.universestay.project.common.PageHandler;
+import com.universestay.project.common.SearchCondition;
 import com.universestay.project.admin.service.UserManagementService;
 import com.universestay.project.user.dto.UserDto;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,13 +18,13 @@ import java.util.Map;
 // TODO: 2023-11-28     css 태그에 직접 입력한거 다듬기
 
 @Controller
-@RequestMapping("/admin")
+@RequestMapping("/admin/user")
 public class UserManagementController {
     @Autowired
     UserManagementService userManagementService;
 
     // 유저리스트 페이지 화면 출력
-    @GetMapping("/userlist")
+    @GetMapping("/list")
     public String userList(Model model, SearchCondition sc) {
         try {
             // Mapper의 쿼리문에 전달해 줄 offset과 pageSize를 map에 저장해서 mapper에 넘겨줌
@@ -59,15 +59,14 @@ public class UserManagementController {
 //            throw new Exception("예외 발생");
         } catch (Exception e) {
             e.printStackTrace();
-        } finally {
-            // 예외가 발생했을 때에는 model이 없는 빈 페이지가 출력
-            // 예외 발생 안했을 시에는 정상적인 페이지 출력
-            return "/admin/userList";
         }
+        // 예외가 발생했을 때에는 model이 없는 빈 페이지가 출력
+        // 예외 발생 안했을 시에는 정상적인 페이지 출력
+        return "/admin/userList";
     }
 
     // 유저리스트에서 페이지에서 아이디를 클릭했을 때 해당 유저의 상세 정보를 출력
-    @GetMapping("/userinfo")
+    @GetMapping("/info")
     public String userInfo(String user_id, Model model) {
         // 매개변수로 user_id 받아옴
         // mapper에 user_id 넘겨주고 userDto 받아오기
@@ -77,10 +76,9 @@ public class UserManagementController {
             model.addAttribute("dto", dto);
         } catch (Exception e) {
             e.printStackTrace();
-        } finally {
-            // 예외가 발생했을 때에는 model이 없는 빈 페이지가 출력
-            // 예외 발생 안했을 시에는 정상적인 페이지 출력
-            return "/admin/userInfo";
         }
+        // 예외가 발생했을 때에는 model이 없는 빈 페이지가 출력
+        // 예외 발생 안했을 시에는 정상적인 페이지 출력
+        return "/admin/userInfo";
     }
 }
