+++ src/main/java/com/universestay/project/admin/dao/UserManagementDao.java
@@ -1,6 +1,6 @@
 package com.universestay.project.admin.dao;
 
-import com.universestay.project.admin.dto.SearchCondition;
+import com.universestay.project.common.SearchCondition;
 import com.universestay.project.user.dto.UserDto;
 
 import java.util.List;
