+++ src/main/java/com/universestay/project/common/PageHandler.java
@@ -1,4 +1,4 @@
-package com.universestay.project.admin.dto;
+package com.universestay.project.common;
 
 import org.springframework.web.util.UriComponentsBuilder;
